requirements.txt
Flask==3.1.0
google-genai==1.9.0
python-dotenv==1.0.1
pytest==8.3.5
python-docx==1.1.2
xhtml2pdf==0.2.16
markdown==3.7
